% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vat.R
\name{get_vat}
\alias{get_vat}
\title{get_vat}
\usage{
get_vat(
  key = Sys.getenv("apelex_api_key"),
  test = FALSE,
  report = c("CountyGeoStrata", "CountyPartyStrata", "CountyPartyStrataEstimates",
    "CountyGeoStrataEstimates", "CountyPartyStatewideEstimates",
    "CountyGeoStatewideEstimates")
)
}
\arguments{
\item{key}{Supplied API key}

\item{report}{The type of VAT report to get data from. Options include: CountyGeoStrata, CountyPartyStrataEstimates, CountyGeoStrataEstimates, CountyPartyStatewideEstimates, CountyGeoStatewideEstimates}
}
\value{
Returns a tidy dataframe of VAT results
}
\description{
Function to get vote count data aggregated by strata, geographic region, or party estimates from the AP voter analysis tool
}
